Move SingleColumnWork into a component
